Refactored system packages Added skeleton for TaskScheduler
